[ADF-123] [BE] Cập nhập logic DBContext trong Repository

+++ Repositories/Implement/FeedBackRepository.cs
@@ -3,39 +3,67 @@ using Microsoft.EntityFrameworkCore;
 
 namespace Repositories
 {
-    public class FeedBackRepository : SingletonBase<FeedBackRepository>, IFeedBackRepository
+    public class FeedBackRepository : IFeedBackRepository
     {
-        public FeedBackRepository() { }
         public async Task Create(Feedback feedback)
         {
-            await context.Feedbacks.AddAsync(feedback);
-            await context.SaveChangesAsync();
+            using (var context = new PostgresContext())
+            {
+                await context.Feedbacks.AddAsync(feedback);
+                await context.SaveChangesAsync();
+            }
         }
 
         public async Task Delete(Guid id)
         {
-            var feedback = await context.Feedbacks.SingleOrDefaultAsync(i => i.Id == id);
-            context.Feedbacks.Remove(feedback);
-            await context.SaveChangesAsync();
+            using (var context = new PostgresContext())
+            {
+                var feedback = await context.Feedbacks
+                    .SingleOrDefaultAsync(i => i.Id == id);
+
+                if (feedback != null)
+                {
+                    context.Feedbacks.Remove(feedback);
+                    await context.SaveChangesAsync();
+                }
+            }
         }
 
         public async Task<IEnumerable<Feedback>> GetAlls()
         {
-            var feedbacks = await context.Feedbacks.ToListAsync();
-            return feedbacks;
+            using (var context = new PostgresContext())
+            {
+                var feedbacks = await context.Feedbacks.ToListAsync();
+                return feedbacks;
+            }
         }
 
         public async Task<Feedback> GetById(Guid id)
         {
-
-            var feedback = await context.Feedbacks.SingleOrDefaultAsync(i => i.Id == id);
-            return feedback;
+            using (var context = new PostgresContext())
+            {
+                var feedback = await context.Feedbacks
+                    .SingleOrDefaultAsync(i => i.Id == id);
+                return feedback;
+            }
         }
 
         public async Task Update(Feedback feedback)
         {
-            context.Entry<Feedback>(feedback).State = EntityState.Modified;
-            await context.SaveChangesAsync();
+            using (var context = new PostgresContext())
+            {
+                var localFeedback = context.Set<Feedback>()
+                    .Local
+                    .FirstOrDefault(entry => entry.Id.Equals(feedback.Id));
+
+                if (localFeedback != null)
+                {
+                    context.Entry(localFeedback).State = EntityState.Detached;
+                }
+
+                context.Entry(feedback).State = EntityState.Modified;
+                await context.SaveChangesAsync();
+            }
         }
     }
 }
